- Refactoring to support reading from either geopackage extension  - reading triggers a detection mechanism, which instantiates the appropriated class using polymorphic behavior.  - writing is currently supported for the qgis_geopackage extension *only*.
