+++ qgis_plugin/qgpkg/qgis_geopackage.py
@@ -27,11 +27,14 @@ import resources
 import os
 import tempfile
 import logging
+import sqlite3
 
 from qgpkg import QGpkg
-
+from qgpkg_owc import QGpkg_owc
+from qgpkg_qgis import QGpkg_qgis
 
 message_bar = None
+
 
 def qlog(lvl, msg, *args, **kwargs):
     msg_level = 2  # TODO:
@@ -45,7 +48,6 @@ def qlog(lvl, msg, *args, **kwargs):
         message_bar.pushMessage(
             msg, level=msg_level)
 
-
 class QgisGeopackage(QObject):
     """QGIS Plugin Implementation."""
 
@@ -56,6 +58,10 @@ class QgisGeopackage(QObject):
         self.toolbar = self.iface.addToolBar(u'Qgis Geopackage')
         self.toolbar.setObjectName(u'Qgis Geopackage')
         message_bar = self.iface.messageBar()
+        self._log = qlog
+
+    def log(self, lvl, msg, *args, **kwargs):
+        self._log(lvl, msg, *args, **kwargs)
 
     def initGui(self):
         pluginPath = QFileInfo(os.path.realpath(__file__)).path()
@@ -77,7 +83,7 @@ class QgisGeopackage(QObject):
         self.actionWrite.setWhatsThis(self.tr(u"Write project in GeoPackage"))
         self.iface.addPluginToMenu("&Qgis Geopackage", self.actionWrite)
         self.toolbar.addAction(self.actionWrite)
-        self.actionWrite.setEnabled(False)
+        # self.actionWrite.setEnabled(False)
         QObject.connect(self.actionWrite, SIGNAL("triggered()"), self.write)
 
         self.actionRead = QAction(
@@ -110,16 +116,72 @@ class QgisGeopackage(QObject):
         else:
             project_path = project.fileName()
 
-        gpkg = QGpkg(project_path, qlog)
+        gpkg = QGpkg_qgis(project_path, qlog)
         gpkg.write(project_path)
 
         if tmpfile:
             os.remove(tmpfile)
 
     def read(self):
+        """Reads a geopackage file polymorphically, according to the auto-detected extension """
+
         gpkg_path = QFileDialog.getOpenFileName(
             self.iface.mainWindow(), self.tr(u"Choose GeoPackage..."),
             None, "GeoPackage (*.gpkg)")
         if gpkg_path:
-            gpkg = QGpkg(gpkg_path, qlog)
-            project_path = gpkg.read(gpkg_path)
+            gpkg = self.detect_gpkg_extension(gpkg_path, qlog)
+            if gpkg is not None:
+                project_path = gpkg.read(gpkg_path)
+            else:
+                self.log(logging.ERROR,
+                         u"We were unable to read this geopackage file")
+                return
+
+    def detect_gpkg_extension(self, gpkg_path, qlog):
+        """Detects which geopackage extension we need to load (if any)
+        and instantiates the subclass of qgpkg, according to it.
+
+        Args:
+            gpkg_path: The path of the gpkg file.
+            qlog: log function
+        Returns:
+            An handle to the instantiated subclass
+        """
+
+        if self.checkIfTableExists("qgis_projects", gpkg_path) is True:
+            gpkg = QGpkg_qgis(gpkg_path, qlog)
+        elif self.checkIfTableExists("owc_context", gpkg_path) is True:
+            gpkg = QGpkg_owc(gpkg_path, qlog)
+        else:
+            self.log(logging.ERROR,
+                     u"Sorry: we did not find a valid geopackage extension whithin this geopackage." +
+                     u"\n Supported extensions include owc_geopackage and qgis_geopackage")
+            return
+
+        return gpkg
+
+    def checkIfTableExists(self, table_name, gpkg_path):
+        """Check if a table with a given name,
+        exists in a sqlite3 database
+
+        Args:
+            table_name: The table we are looking for.
+            gpkg_path: The path of the gpkg file.
+        """
+
+        conn = sqlite3.connect(gpkg_path)
+        c = conn.cursor()
+
+        try:
+            c.execute("""
+                SELECT name
+                FROM sqlite_master
+                WHERE type='table' AND name=?;
+            """, (table_name,))
+            ret = bool(c.fetchone())
+            conn.close()
+            return ret
+
+        except sqlite3.OperationalError:
+            conn.close()
+            self.log(logging.ERROR, u"Table '" + table_name + u"' does not seem to exist in the database.")
